requirements.txt
argparse
requests
plexapi
pyaml
